+++ SpaceWarp/UI/Console/SpaceWarpConsoleLogListener.cs
@@ -1,8 +1,8 @@
-﻿using System;	
-using System.Collections.Generic;	
-using BepInEx.Logging;	
+﻿using System;
+using System.Collections.Generic;
+using BepInEx.Logging;
 
-namespace SpaceWarp.UI.Debug;	
+namespace SpaceWarp.UI.Console;	
 
 public sealed class SpaceWarpConsoleLogListener : ILogListener	
 {
